Improve console output, change 1 billion to float

+++ message_queue.c
@@ -57,9 +57,9 @@ void update_termlog(int msgqid, struct termlog* sbuf) {
 }
 
 void increment_sysclock(struct sysclock* sysclock, int increment) {
-    const int ONE_BILLION = 1000000000;
+    const float ONE_BILLION = 1000000000;
     sysclock->clock.nanoseconds += increment;
-    if ( (sysclock->clock.nanoseconds / ONE_BILLION) >= 1) {
+    if ( (sysclock->clock.nanoseconds / ONE_BILLION) >= 1.0) {
         sysclock->clock.seconds += 1;
         sysclock->clock.nanoseconds -= ONE_BILLION;
     }
